Added: - Support of ArduinoIDE 1.8.x, - Support of user-defined configuration parameters. The level of wakeup via INT1 can be changed now using zunoSetWUOptions. - Added ability define battery levele for empty and full battery (use ZUNO_SETUP_BATTERY_LEVELS macro) - Added ability to use user-defined battery level handler (ZUNO_SETUP_BATTERY_HANDLER). - User can read battery monitor values directly using AnalogRead(BATTERY). -Added WUT setings in FLIRS mode (zunoSetBeamCountWU). Fixed:- channel handlers problems in FLIRS. In short words there are a lot of changes! The firts beta of 2.1.1 is ready.

+++ hardware/arduino/zuno/cores/zuno/Print.h
@@ -37,24 +37,24 @@ class Print
     
     size_t print(char *);
     size_t print(char);
-    size_t print(unsigned char, int = DEC);
-    size_t print(int, int = DEC);
-    size_t print(unsigned int, int = DEC);
-    size_t print(long, int = DEC);
-    size_t print(unsigned long, int = DEC);
-    size_t print(float, int = 2);
-    size_t print(double, int = 2);
+    size_t print(unsigned char, uint8_t = DEC);
+    size_t print(int, uint8_t = DEC);
+    size_t print(unsigned int, uint8_t = DEC);
+    size_t print(long, uint8_t = DEC);
+    size_t print(unsigned long, uint8_t = DEC);
+    size_t print(float, uint8_t = 2);
+    size_t print(double, uint8_t = 2);
 
 
     size_t println(char *);
     size_t println(char);
-    size_t println(unsigned char, int = DEC);
-    size_t println(int, int = DEC);
-    size_t println(unsigned int, int = DEC);
-    size_t println(long, int = DEC);
-    size_t println(unsigned long, int = DEC);
-    size_t println(float, int = 2);
-    size_t println(double, int = 2);
+    size_t println(unsigned char, uint8_t = DEC);
+    size_t println(int, uint8_t = DEC);
+    size_t println(unsigned int, uint8_t = DEC);
+    size_t println(long, uint8_t = DEC);
+    size_t println(unsigned long, uint8_t = DEC);
+    size_t println(float, uint8_t = 2);
+    size_t println(double, uint8_t = 2);
     
     size_t println(void);
